Update calculator.py
Stopped the sum displaying outside of the window. I need to check if buttons have been clicked because there's a flicker that occurs without it. It's because I'm removing a frame, then adding it back to the window afterwards. It's a hacky way of solving the problem but it works.